Work toward migrating mappers to new framework:  - GetOrganizationTreeDTO  - GetOrganizationsByIds
Deleted deprecated code:
 - OrganizationHierarchyCacheLoader

Most of this work is splitting apart GetOrganizationsByIds.  It had three public methods - one to get all organizations, one to get a specific one, and one to get a list.  Now, it just implements the list one.  Any use of the single org method was replaced with a call to a list with one org.  Calls to get all orgs (org tree) was replaced with a call to a new mapper that gets a list of all organization ids, and then a call to GetOrganizationsByIds with those ids.

The app is currently wired to use the database mapper to get organizations by ids, and the organization tree is not currently cached.  My next checkin will mainly be a config change that will wire up a new DecoratedPartialResponseDomainMapper for each the org tree loader and the get organizations by id mapper.